agent/requirements.txt
python-dotenv
chromadb==1.1.1
langchain>=0.1.0
langchain-chroma
langchain-community
langchain-core
langchain-ollama
langgraph
langsmith
streamlit
fastapi>=0.100
uvicorn>=0.24
requests>=2.31
pydantic>=2.5
requests>=2.32
deepagents